:sparkles: :: [#289] LectureDetailSetting / 담당 교수 선택 Popup -> BottomSheet
